// packages/shared/src/firebase/models/encounter.model.ts
import type { DocumentData } from 'firebase-admin/firestore';
import type { PartyStatus } from './signup.model.ts';

export interface EncounterDocument extends DocumentData {
  name: string;
  description: string;
  active: boolean;
  mode?: 'legacy' | 'ultimate' | 'savage';
  emoji?: string;
  fflogsEncounterIds?: number[];
  progPartyThreshold?: string;
  clearPartyThreshold?: string;
}

export interface ProgPointDocument extends DocumentData {
  id: string; // Short name used for public presentation (Google Sheets, etc.)
  label: string; // Long name used for Discord presentation (signup messages, embeds, etc.)
  partyStatus: PartyStatus;
  order: number;
  active: boolean;
}